+++ lib_test/global_extended_headers_test.ml
@@ -2,7 +2,7 @@ let level = Tar.Header.Ustar
 
 module Writer = struct
   type out_channel = Stdlib.out_channel
-  type 'a t = 'a
+  type 'a io = 'a
   let really_write oc cs =
     let str = Cstruct.to_string cs in
     output_string oc str
@@ -16,7 +16,7 @@ module HW = Tar.HeaderWriter
 
 module Reader = struct
   type in_channel = Stdlib.in_channel
-  type 'a t = 'a
+  type 'a io = 'a
   let really_read ic cs =
     let len = Cstruct.length cs in
     let buf = Bytes.create len in
@@ -58,7 +58,8 @@ let use_global_extended_headers _test_ctxt =
   let cout = open_out_bin "test.tar" in
   let g0 = make_extended 1000 in
   let hdr, f = make_file () in
-  HW.write ~level ~global:g0 hdr cout;
+  HW.write_global_extended_header g0 cout;
+  HW.write ~level hdr cout;
   f cout;
   let hdr, f = make_file () in
   let hdr = { hdr with Tar.Header.extended = Some (make_extended 2000) } in
@@ -69,7 +70,8 @@ let use_global_extended_headers _test_ctxt =
   f cout;
   let g1 = make_extended 3000 in
   let hdr, f = make_file () in
-  HW.write ~level ~global:g1 hdr cout;
+  HW.write_global_extended_header g1 cout;
+  HW.write ~level hdr cout;
   f cout;
   Writer.really_write cout Tar.Header.zero_block;
   Writer.really_write cout Tar.Header.zero_block;
@@ -81,7 +83,7 @@ let use_global_extended_headers _test_ctxt =
     let pp ppf hdr = Fmt.pf ppf "%s" (Tar.Header.Extended.to_detailed_string hdr) in
     Alcotest.testable (fun ppf hdr -> Fmt.pf ppf "%a" Fmt.(option pp) hdr) ( = )
   in
-  ( match HR.read ~level ~global:!global cin with
+  ( match HR.read ~global:!global cin with
     | Ok (hdr, global') ->
        Alcotest.check header "expected global header" (Some g0) global';
        global := global';
@@ -89,7 +91,7 @@ let use_global_extended_headers _test_ctxt =
        let to_skip = Tar.Header.(Int64.to_int (to_sectors hdr) * length) in
        Reader.skip cin to_skip;
     | Error `Eof -> failwith "Couldn't read header" );
-  ( match HR.read ~level ~global:!global cin with
+  ( match HR.read ~global:!global cin with
     | Ok (hdr, global') ->
        Alcotest.check header "expected global header" (Some g0) global';
        global := global';
@@ -97,7 +99,7 @@ let use_global_extended_headers _test_ctxt =
        let to_skip = Tar.Header.(Int64.to_int (to_sectors hdr) * length) in
        Reader.skip cin to_skip;
     | Error `Eof -> failwith "Couldn't read header" );
-  ( match HR.read ~level ~global:!global cin with
+  ( match HR.read ~global:!global cin with
     | Ok (hdr, global') ->
        Alcotest.check header "expected global header" (Some g0) global';
        global := global';
@@ -105,7 +107,7 @@ let use_global_extended_headers _test_ctxt =
        let to_skip = Tar.Header.(Int64.to_int (to_sectors hdr) * length) in
        Reader.skip cin to_skip;
     | Error `Eof -> failwith "Couldn't read header" );
-  ( match HR.read ~level ~global:!global cin with
+  ( match HR.read ~global:!global cin with
     | Ok (hdr, global') ->
        Alcotest.check header "expected global header" (Some g1) global';
        global := global';
@@ -113,7 +115,7 @@ let use_global_extended_headers _test_ctxt =
        let to_skip = Tar.Header.(Int64.to_int (to_sectors hdr) * length) in
        Reader.skip cin to_skip;
     | Error `Eof -> failwith "Couldn't read header" );
-  ( match HR.read ~level ~global:!global cin with
+  ( match HR.read ~global:!global cin with
     | Ok _ -> failwith "Should have found EOF"
     | Error `Eof -> () );
   ()